requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
Django==4.2.6
django-cors-headers==4.2.0
djangorestframework==3.14.0
exceptiongroup==1.1.3
fake-useragent==1.3.0
frozenlist==1.4.0
h11==0.14.0
idna==3.4
joblib==1.3.2
multidict==6.0.4
nltk==3.8.1
numpy==1.26.0
openai==0.28.1
outcome==1.2.0
pandas==2.1.1
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
regex==2023.10.3
requests==2.31.0
scikit-learn==1.3.1
scipy==1.11.3
selenium==4.14.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
threadpoolctl==3.2.0
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
wsproto==1.2.0
yarl==1.9.2